Enable grub serial console and increase timeout
On Azure, by default the root account is locked and sudo is enabled to elevate privileges from a non-root user.  This means that the emergency shell is of no use on CentOS (and similar variants) because sulogin on RHEL/CentOS requires a root password to open a rescue shell.

Therefore, in order to recover a broken system via serial console one must edit the kernel parameters to boot into single user mode.  This requires that the grub menu is operating on the serial console and that the timeout is of sufficient size to allow one interrupt system boot.